troubleshooting/Part_5_2/requirements.txt
aws-cdk-lib==2.126.0
constructs>=10.0.0,<11.0.0
aws-cdk.lambda-layer-kubectl-v28
boto3